Start of bConomy...
Basically idea is...
on player login, do sql query to get information. Store information and use that information to do add/subtract.
Also when a transaction is done, add to save, but that can be done in a thread, so no lag there.

Also would be nice that, we don't remove a players account on logout, but keep it for 5mins~ just in case they come back on?

git-svn-id: https://subversion.assembla.com/svn/bconomy@2 59879121-d2b3-4da0-87af-382deb114e0d